// src/api/model/testInfo.ts
/**
 * Generated by orval v6.26.0 🍺
 * Do not edit manually.
 * student-hub-api
 * Api Documentation for StudentHub API
 * OpenAPI spec version: 0.0
 */
import type { TestInfoExercisesItem } from './testInfoExercisesItem';
import type { TestFeedback } from './testFeedback';

export interface TestInfo {
  bestGrade?: number;
  bestSubTopic?: string;
  createdAt?: string;
  difficulty?: number;
  exercises?: TestInfoExercisesItem[];
  isExam?: boolean;
  numberOfExercises?: number;
  questionType?: number;
  testEventId?: number;
  testId?: number;
  testResult?: TestFeedback;
  timeElapsed?: number;
  worstSubTopic?: string;
}