// src/api/model/exerciseBasicInfo.ts
/**
 * Generated by orval v6.26.0 🍺
 * Do not edit manually.
 * student-hub-api
 * Api Documentation for StudentHub API
 * OpenAPI spec version: 0.0
 */

export interface ExerciseBasicInfo {
  difficulty: number;
  exerciseId: number;
  isAnswered: boolean;
  isExam?: boolean;
  isMultipleChoice: boolean;
}